3619: Rename SmartAttributeEditorManager files
