+++ common/src/View/SmartPropertyEditorManager.h
